A change in the message display mechanism
* When using the Message component, a problem was found: when it was
necessary to call messages sequentially, the behavior did not match
the expected one. Therefore, it became necessary to change the
messageReducer and the Message component to work with a message queue
instead of a single message. Now you only need to enqueue a
message using the appropriate action, the display is performed
automatically.
* Corresponding changes have been made to the components that
currently use the messageReducer: HeaderActions, Footer, Login.