Changing the _global.scss file
An additional style has been added for the [#root > div > div] element
which has a child component with the 'overlay' class name. In fact,
this is a div element between the Header and Footer. When the Loader
component is displayed, the parent div should have 'flex' as the
display property.