Adding the ability to add items to the cart
The MainBestProduct and ProductCard components have been changed.
Now the MainBestProduct component uses cartApi.getCart endpoint as a
promise. It receives the status of the cart after loading, then after
logging out or when needRefetch flag is set in any ProductCard
instance. In the ProductCard, the user can add some item to the cart,
and then the server will send a response with the current status of the
cart, which is applied in the MainBestProduct component. The
needRefetch flag is set when there is no required field in the serponse
after adding the item, and this is necessary to synchronize the status
of the cart.