Adding a new useCartSync hook
It is used to sync the user's cart. If some items were added to the cart
while the user was not logged in, an attempt will be made to sync the
anonymous cart with the cart stored on the server after the user is
logged in. Unfortunately, there is no endpoint that can be used to
update the entire cart in one request, so it is necessary to use the
Promise.all method to send multiple update requests, which takes more
time.
The algorithm is as follow:
   1. If an item is not in a remote cart, it will be added.
   2. If the quantity of an item in an anonymous cart is less than
   the quantity of the same item in a remote cart, it will not
   be applied, otherwise a new quantity will be applied.
   3. If the synchronization attempt is successful, a success
   message will appear and the cartToSync will be cleared.
   4. If after three unsuccessful attempts the cart is not
   syncronized, an error message will appear and cartToSync will
   be cleared.